fix flyer

+++ crop_insta.sh
@@ -1,3 +1,12 @@
 thu=$(date +"%Y-%m-%d" -d "$(date +"%Y%m%d + 4 days - %w days")")
-convert $1 -crop 1080x1920+0+223  images/ttt/$thu.png
-rm $1
+
+height=$(magick identify -format "%H\n" $1)
+
+if [ $height -eq 1080 ]; then
+        echo "height is correct, cropping"
+        convert $1 -crop 1080x1920+0+223  images/ttt/$thu.png
+        rm $1
+else
+        echo $height "not correct"
+fi
+
